move object extraction into CacheRead
This seems cleaner than having somebody else manage object extraction.
